Renamed mapping namespace to mappers
