+++ Dfe.Data.SearchPrototype/Web/Program.cs
@@ -7,12 +7,12 @@ using GovUk.Frontend.AspNetCore;
 using Azure.Search.Documents.Models;
 using Azure;
 using DfE.Data.ComponentLibrary.CrossCuttingConcerns.Mapping;
-using Dfe.Data.SearchPrototype.Infrastructure.Mapping;
+using Dfe.Data.SearchPrototype.Infrastructure.Mappers;
 using Azure.Search.Documents;
 using Dfe.Data.SearchPrototype.Infrastructure.Options.Mapping;
 using Dfe.Data.SearchPrototype.Infrastructure.Options;
 using Dfe.Data.SearchPrototype.Web.Models;
-using Dfe.Data.SearchPrototype.Web.Mapping;
+using Dfe.Data.SearchPrototype.Web.Mappers;
 using Infrastructure = Dfe.Data.SearchPrototype.Infrastructure;
 using DfE.Data.ComponentLibrary.CrossCuttingConcerns.Json.Serialisation;
 using Search = Dfe.Data.SearchPrototype.Search;
